Create Iris-ZWave-Repeater.src
Check the health of the Z-Wave repeaters within your Iris Smart Plugs!   This driver allows you to keep tabs on whether the Z-Wave portion of the device is working and checks it every 30 minutes.

After installing this driver and selecting it for each device, it is important to click the Configure() button to that the association group can be set allowing the plug to communicate its ownstatus with the hub.